NP-17160-embargo-detector, code review
